account_profiles_controllerのディレクトリ変更、AccountProfileModelのtypoの修正
