+++ config/routes.rb
@@ -17,8 +17,8 @@ Rails.application.routes.draw do
 
   namespace :v1 do
     resources :accounts, only: %i[index show destroy] do
-      resources :account_profile, only: %i[index show destroy]
-    end 
+      resources :account_profiles, only: %i[index create show update destroy]
+    end
     resources :occupation_main_categories, only: :index
     resources :industry_categories, only: :index
     resources :prefectures, only: :index
