--- spec/requests/account_profiles_request_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "AccountProfiles", type: :request do
-
-end
