+++ spec/requests/v1/account_profiles_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "V1::AccountProfiles", type: :request do
+
+end
